# lib/api-zod/src/generated/types/atsTrendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface AtsTrendPoint {
  date: string;
  score: number;
  filename: string;
}