# lib/api-zod/src/generated/types/interviewSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewSession {
  id: number;
  role: string;
  experienceLevel: string;
  firstQuestion: string;
  /** @nullable */
  createdAt?: string | null;
}