# lib/api-zod/src/generated/types/interviewSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewSummary {
  id: number;
  role: string;
  experienceLevel: string;
  /** @nullable */
  overallScore?: number | null;
  questionsAnswered?: number;
  createdAt: string;
}