# lib/api-zod/src/generated/types/jobMatchRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface JobMatchRecord {
  id: number;
  jobTitle: string;
  matchPercentage: number;
  createdAt: string;
}