# lib/api-zod/src/generated/types/jobMatchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface JobMatchResult {
  id?: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  matchPercentage: number;
  skillGaps: string[];
  improvementSuggestions: string;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  createdAt?: string | null;
}